+++ src/SubmarineFree.cpp
@@ -73,6 +73,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelVM104);
 	p->addModel(modelVM201);
 	p->addModel(modelVM202);
+	p->addModel(modelVM204);
 
 	p->addModel(modelWK101);
 	p->addModel(modelWK205);