+++ changelog.md
@@ -21,6 +21,7 @@
 - 12-Nov-2021 Add expansion capabilities to VM-1xx and VM-2xx devices
 - 14-Nov-2021 Use keywords to make library searching more effective
 - 14-Nov-2021 Add VM-104 Quad digital VU-meter
+- 19-Nov-2021 Add VM-204 Quad vintage VU-meter
 
 ## v1.1.10
 - 10-Dec-2020 Update images in manual